refactor(BoardGenerator): 보드 생성 부분 클래스 분리
